requirements.txt
boto3
numpy
pandas
ipywidgets
tqdm
Pillow
jupyterlab_widgets